Display of graphes.
Work started on random graph generation. Not working, displays no edges.

Update random_pick to work without the need to construct
variables of the output type. Copy only.

Creating edges without a cost, for having temp edges. Create a
temp edge with only start and end, to find if such an edge exists
in a graph.