Generic priority queue.
Generic data type for the priority queue elements. It was int, but for Kruskal
we need the ability to save edges in the queue.

Prioqueue's prio_set function needs to reuse the Data field that was present
earlier, and not the one passed as parameter for comparison. This is because the
data is saved as reference, and the parameter passed for comparison may be a
local variable.